Refactoring module lookup for use with VendorJob
