+++ spec/services/metrc_service/batch_spec.rb
@@ -103,66 +103,4 @@ RSpec.describe MetrcService::Batch do
       it { is_expected.to be_nil }
     end
   end
-
-  context 'module lookup' do
-    let(:task) { create(:task, integration: integration) }
-    let(:instance) { described_class.new(ctx, integration, nil, task) }
-    let(:seeding_unit) { double(:seeding_unit, name: seeding_unit_name) }
-    before do
-      allow(instance)
-        .to receive(:seeding_unit)
-        .and_return(seeding_unit)
-    end
-
-    context '#module_name_for_seeding_unit' do
-      subject { instance.send(:module_name_for_seeding_unit) }
-
-      context 'plants_barcoded' do
-        let(:seeding_unit_name) { 'Plants (Barcoded)' }
-        it { is_expected.to eq('plant') }
-      end
-
-      context 'package' do
-        let(:seeding_unit_name) { 'Package' }
-        it { is_expected.to eq('package') }
-      end
-
-      context 'testing_package' do
-        let(:seeding_unit_name) { 'Testing Package' }
-        it { is_expected.to eq('package') }
-      end
-
-      context 'sales_order' do
-        let(:seeding_unit_name) { 'Sales Order' }
-        it { is_expected.to eq('sales_order') }
-      end
-    end
-
-    context '#module_for_completion' do
-      let(:action_type) { 'start' }
-      let(:completion) { double(:completion, action_type: action_type) }
-      subject { instance.send(:module_for_completion, completion) }
-
-      context 'plants_barcoded' do
-        let(:seeding_unit_name) { 'Plants (Barcoded)' }
-        it { is_expected.to eq(MetrcService::Plant::Start) }
-      end
-
-      context 'package' do
-        let(:seeding_unit_name) { 'Package' }
-        it { is_expected.to eq(MetrcService::Package::Start) }
-      end
-
-      context 'testing_package' do
-        let(:seeding_unit_name) { 'Testing Package' }
-        it { is_expected.to eq(MetrcService::Package::Start) }
-      end
-
-      context 'sales_order' do
-        let(:action_type) { 'discard' }
-        let(:seeding_unit_name) { 'Sales Order' }
-        it { is_expected.to eq(MetrcService::SalesOrder::Discard) }
-      end
-    end
-  end
 end
